% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDB.R
\name{findCellTypes}
\alias{findCellTypes}
\title{Greedy search for celltypes within a specific species}
\usage{
findCellTypes(celltype, species)
}
\arguments{
\item{celltype}{partial cell type name}

\item{species}{name of species to look into}
}
\value{
vector of celltypes
}
\description{
Sometimes the name of a cell-type is only partially known, for these circumstances, this function was designed.
Using grep all celltypes which contain the "celltype" variable are returned.
}
