% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDB.R
\name{getCelltypeEnrichment}
\alias{getCelltypeEnrichment}
\title{Returns celltype predictions for the provided list of genes}
\usage{
getCelltypeEnrichment(genes, species = NULL,
  server = "http://markers.blebli.de")
}
\arguments{
\item{genes}{Vector with genes (Ensembl IDs)}

\item{species}{Ensembl name of species (hsapiens_gene_ensembl), necessary for celltype prediction including orthologues}

\item{server}{URL of the server hosting the markerDB website.}
}
\value{
data.frame with predicted celltypes and p-values.
}
\description{
Returns celltype predictions for the provided list of genes
}
