% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDB.R
\name{getSpecies}
\alias{getSpecies}
\title{Get all species in MarkerDB}
\usage{
getSpecies(server = "http://markers.blebli.de")
}
\arguments{
\item{server}{URL of the server hosting the markerDB website.}
}
\value{
character vector of species in MarkerDB
}
\description{
Get all species in MarkerDB
}
