% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDB.R
\name{getSpeciesInfo}
\alias{getSpeciesInfo}
\title{Get informations of a species in MarkerDB}
\usage{
getSpeciesInfo(species, server = "http://markers.blebli.de")
}
\arguments{
\item{species}{Character variable of species name}

\item{server}{URL of the server hosting the markerDB website.}
}
\value{
List of species annotations
}
\description{
This function helps to get information of a specieid species in MarkerDB.
The user not only gets the names of all celltypes known within in this species,
but also its latin and ENSEMBL name.
}
