% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDB.R
\name{getMarkerGenes}
\alias{getMarkerGenes}
\title{Get all markers for a specific cell type within a species}
\usage{
getMarkerGenes(celltype, species, field = c("marker_accession",
  "marker_name"), positive = TRUE, filter_list = list())
}
\arguments{
\item{celltype}{name of celltype}

\item{species}{name of species (as can be found using getSpecies function)}

\item{field}{can be either "marker_accession" to get GeneIDs or "marker_name" to return gene names}

\item{positive}{True by default. Only return positive markers}

\item{filter_list}{Additional filters in list format (i.e: list(stage = "p7")).
Wildcards can be added by using a "\%" sign.
So, if all embryonic mouse markers are wanted, a list like list(stage="E\%") can be used.}
}
\value{
character vector of marker genes
}
\description{
This functions gives a vector of annotated markers of a cell type within a predefined species.
}
