% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDB.R
\name{getSpeciesCellTypeInfo}
\alias{getSpeciesCellTypeInfo}
\title{Get a table of celltype informations}
\usage{
getSpeciesCellTypeInfo(celltype, species,
  server = "http://markers.blebli.de")
}
\arguments{
\item{celltype}{name of celltype}

\item{species}{name of species}

\item{server}{URL of the server hosting the markerDB website.}
}
\value{
data.frame of informations on all markers of a specific cell type.
}
\description{
To get a data.frame which holds all informations of a cell type within a specified species, use this function.
}
